requirements.txt
pytesseract
Pillow
scikit-learn
joblib
numpy
gunicorn==20.1.0
Flask
flask-cors
requests
beautifulsoup4
python-dotenv
google-generativeai
